requirements.txt
tensorflow
numpy
matplotlib
seaborn
scikit-learn
streamlit
Pillow